Merged PR 9486: airlock workspace version
The Airlock workspace pulls it's terraform from the base workspace, in a previous PR the version of the airlock workspace should have been changed and was omitted.  This fixes it.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a version update in a workspace configuration file.

#### PR Summary
The pull request bumps the airlock import review workspace version from 0.14.9 to 0.14.10 to address linked work items ensuring that MCAPs policies do not break things overnight.
- `templates/workspaces/airlock-import-review/porter.yaml`: Version number updated from 0.14.9 to 0.14.10.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #60086