Merged PR 9505: Fixing workspace creation bug
Fixing workspace creation bug.

This was due to an `*` value being used for a field that doesn't support it. Fix is to change the field to from `destination_address_prefixes` to `destination_address_prefix` and the value to `"*"`.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix addressing workspace creation issues.

#### PR Summary
This PR fixes a bug that prevented the creation of workspaces by updating a misconfigured network security rule and bumping the workspace base version.
- `/templates/workspaces/base/terraform/network/security.tf`: Modified the outbound Windows activation rule by replacing `destination_address_prefixes` with `destination_address_prefix`.
- `/templates/workspaces/base/porter.yaml`: Updated the version from 2.5.10 to 2.5.11.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #60116